Moved output functions to backend
